Add support for django with multiple databases
This allows to switch schema for all connections rather than the only default database.

The changes assume that all connections are always on the same schema as the first connection (`'default'`), meaning the tasks are not switching schema during their execution. Thus that we don't need to check all connections whether they are on a different schema or not.

The lines where the assumption is made:
- app.py:33
- app.py:39
- app.py:42
- app.py:46
- app.py:65